requirements.txt
Flask>=2.3
google-genai>=1.0.0
